Adding 1st, 3rd & 4th multiplexers.
They have the same design in MIPS Monocycle, i.e., two 32 bits inputs and a single-bit input to select an entrance, with the 32 bits output. The 2nd multiplexer, however, have two 5 bits inputs, a single-bit input and a 5 bits output. Also changing the folders to align with this usage of the same code for three multiplexers.